+++ src/app/app.component.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
-import { HeaderModule } from './components/header/header.module';
+import { HeaderModule } from './shared/components/header/header.module';
 
 describe('AppComponent', () => {
   beforeEach(async () => {