+++ src/app/guards/auth.interceptor.ts
@@ -1,4 +1,4 @@
-import { LocalStorageService } from './../services/local-storage.service';
+import { LocalStorageService } from './../shared/services/local-storage.service';
 import { Injectable } from '@angular/core';
 import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor } from '@angular/common/http';
 import { Observable } from 'rxjs';